Move functions dealing with archived sample sheet to housekeeper storage func
